Add support for changing the outbound codec on a SIP call using a dialplan variable.
This adds a dialplan variable (SIP_CODEC_OUTBOUND) which controls
the codec offered for an outgoing SIP call. This is much like the
SIP_CODEC dialplan variable and has the same restrictions. The codec
set must be one that is configured for the call.

(closes issue #13243)
Reported by: samdell3
Patches:
      13243.diff uploaded by file (license 11)


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@186624 f38db490-d61c-443f-a65b-d21fe96a405b