Define our desires for POSIX and X/OPEN API features properly.
Based on a post on the gcc-help mailing list and some subsequent reading,
we can increase our portability to various platforms by directly defining
the POSIX and X/OPEN API feature sets we wish to have available. This patch
does that, and also includes a double-check to ensure that the system
we are compiling on can actually provide the requested feature sets.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@212463 f38db490-d61c-443f-a65b-d21fe96a405b