Added support for postgres database retry query on disconnection to res_config_pgsql
If your postgres connection died suddenly in between res_config_pgsql
queries, the next query will fail because the query is executed on a
disconnected/disconnecting handle.  The query is abandoned and is
returned from in error.

Now we will reconnect and try again if a query was run on a
disconnected connection.

(closes issue #18071)


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@300882 f38db490-d61c-443f-a65b-d21fe96a405b